Create flashgot.xml
FlashGot is made by the same dude that made NoScript. The log file contains the EXACT addresses of files downloaded with it and this should immediately be pushed when ready for the sake of privacy. I currently only have the Firefox on GNU+Linux version and I am going to now try and add the Windows version too.

I tested this with BB 1.4 on GNU+Linux Mint 17 and it cleaned up a little over half a MB.