Create virtualdub-scripts-and-sparse.xml
This is an alternate, more Bolshevik, version of the VirtualDub for Windows cleaner. This cleanerML contains file purge actions that may or may not be what a VirtualDub user wants, and thus, why I left it out of the original virtualdub.xml. For me, I would want them in BleachBit, but I am not "all users". None of the three file types purged are part of a program installation. The VCF, SYL, and JOBS scripts are user-made and are the kinds of things a user may re-use (and other applications may make use of these extensions as well). SPARSE AVI files are user-generated as well and unlikely to be a final/end-project product as they may be very large files.