Windows memory dumps
This is https://github.com/az0/bleachbit/pull/86, except with the Winapp2 entries included. The other PR can be closed. The memory dump entry can be taken out of Winapp2, once this gets merged into CleanerML, as all the other cleaners have it in their own cleaning rules, outside of Winapp2.

This is my first ever trying to understand CleanerML, so there probably is a lot of mistakes.